Merged PR 335: Adds saving of models to a project
This PR targets [Story 504: As a User, I want to save a model design](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_workitems/edit/504)

This PR allows changes where models are saved in the modeling tool.
Prior to this, the model information was kept in the browser's local storage.

The models' source of truth now lies in their respective **models.json** files.

**Rygg** saves the associated between a project and its model, where a project is just a logical construct with no visible manifestation on the file system. It just provides a default directory in which the models are populated.

Tested within in this PR:
- create new project (w/ and w/o other projects existing)
- create new model in project (w/ and w/o other models)
- delete projects (w/ and w/o models)
- delete network from the model view
- reload page in modeling view (to check that local storage holds temp data)
- changing project when there are unsaved changes (even if "changing to same project")
- deleting network when there are unsaved changes

Import of new models were not worked on in this PR.

To test this PR, **rygg** must be on a branch that has **default_directory** and **location** properties for the respective **project** and **model** models (e.g. the [story/504_saving_of_models](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/rygg?version=GBstory%2F504_saving_of_models) branch in the **rygg** repo).

Note: old projects not having a **default_directory** should be deleted and recreated.
Or we can add a context menu to add a **default_directory**

---
--- Update 2020-05-21

To test, please:
- update the **story/504_saving_of_models** branch in the rygg repo,
- remove the old **db.sqlite3** file
- recreate the sqlite3 file by running `./manage migrate`
- make sure you see the following:

```
Running migrations:
  ...
  Applying api.0003_merge_20200521_1703... OK
```
---
--- Update 2020-05-26

A project path now gets created under the users' Document folder.

---
--- Update 2020-05-28

The user now chooses a directory for the project, the models will then be saved under the following scheme:

_<user chosen path>_/_<project name>_/_<model name>_/model.json

Related work items: #504