Merged PR 425: Updates code blocks in notebook view to read only
This PR tackles the problem where it is not obvious that code changes in the notebook view is not saved nor are the underlying layers affected. This PR sets the code blocks in notebook view to read only.

This is a temporary solution, while the ability for code updates in the notebook view to be reflected in the layers' code is being developed.