Merged PR 402: Save workspace changes (in localstorage) until saved or discarded
This PR persists the changes made to the workspace (in localstorage) until the model with the change is saved.

Previously when a user makes a change, and then navigates away and back (or reloads), it would appear that the change has been saved/persisted to disk. This is not true as the change is only saved in the browser. (So reloading the same model would seem that a save didn't work.)