Merged PR 429: [Bug 664] Export and file chooser modal not showing up in notebook view
This PR fixes export and file chooser modal not showing up when in the notebook view.

To test:
- create/select any network
- click on the "Notebook" button on the right side of the toolbar
- wait for code to load
- select "File->Export"
- check that the export modal shows up (as should the file choose when clicking on "Browse")

---
**2020-06-23**

Updates:
- how the save modal is triggered and
- disables the "Save" and "Save As" menu items when in the notebook view:
![image.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/429/attachments/image.png)

I was afraid that the users will think the "**Save [As]**" menu items as actions on the notebook view (for example, using the "**Save As**" instead of "**Export**" when they want a `.ipynb` file).
We're testing to see if the users will accept having both menu items disabled.

Both the "**Save**" and "**Save As**" functions work in notebook view ([saveInNotebookView.mov](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/429/attachments/saveInNotebookView.mov)), so it's only a matter of enabling them in the menubar to get the feature back.

Related work items: #664