Merged PR 461: Switches temporary models storage from LocalStorage to IndexedDB
This PR tackles the problem where the LocalStorage can get full, which causes:
- models to not be viewable in the workspace and
- problems when reloading the page.
<br/>

This PR now saves information in a IndexedDB database called: `perceptilabs-idb`.
The tables used are:
- `ids` (for keeping track of network ids of the current projects) and
- `models` (for the actual model)

Only the models and ids are saved in IndexedDB, which means other functions aren't affected (last used tab, user token). The model and is tables are cleared when logging out.

By having more storage, models are discarded less often.
A side effect of this PR is that there is now less models "blinking" in and out when changing projects.
<br/>

Tested scenarios:
- Login with/without clean LocalStorage
- Creating/importing project
- Creating/importing models
- Reload page while in in workspace and projects view