Merged PR 488: Leaky ReLU and Softmax activations functioning
This is a pull request for the following stories:

#742
#743
## Files to Review
* `backend/perceptilabs/core_new/layers/templates/tf1x.j2`
* `backend/perceptilabs/core_new/layers/utils.py`

## Work Completed
This story includes adding `tf.compat.v1.nn.leaky_relu` and `tf.compat.v1.nn.softmax` to the table of activation functions in `utils.py` and changing the Recurrent DL component to support activation functions. This required changes to the way the layer is defined in `tf1x.j2`. Previously, because it didn't support activation functions, it defaulted to `tanh`.

## Concerns
Please take a look at the Recurrent DL component's changes, specifically regarding the weight and bias variables `W` and `b` and how they're initialized on lines `937-938` in `tf1x.j2`. To test this component, I made an RNN that classifies the mnist data as such:

![Screen Shot 2020-07-09 at 5.32.50 PM.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/488/attachments/Screen%20Shot%202020-07-09%20at%205.32.50%20PM.png)

![Screen Shot 2020-07-09 at 5.37.43 PM.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/488/attachments/Screen%20Shot%202020-07-09%20at%205.37.43%20PM.png)

Related work items: #742, #743