Merged PR 474: Custom popup for all training components
Initial bug was to show correct results in the pop up after the training. For this, Lai worked on the frontend and I made some structural change to endResult dict in core interface as requested by Lai. (This might require a relook to make sure it will support multiple training layers in a model)

During that fix, we came across few more bugs. One of the being, if the number of snapshots being sent to the policies is only 1, then some of the if-conditions aren't allowing the epoch related stats to be updated.

That has been fixed in all the policies and corresponding minor changes are made in all training layers.

Related work items: #707