Merged PR 509: Custom Stopping Conditions
This PR is for the following story: #697

## Files Changed
* `tf1x_classification.j2`, `tf1x_gan.j2`, `tf1x_object_detection.j2`: All of these files have a target accuracy and number of iterations as custom stopping conditions.
* `tf1x_regression.j2`: This file only has number of iterations as a custom stopping condition, because using R-Squared depends on the batch average of the dataset, which means that in some cases training may not even complete one epoch. Because of this, I just removed Target R-Squared a feature for custom stopping.
* `backend/perceptilabs/core_new/layers/utils.py`: This is where the stopping conditions are resolved. I put three conditions: Epochs, Iterations, and Target Accuracy. Every training layer defaults to epochs as the stopping condition, with the other two options selectable for the layers specified in the first point.

## Work Done
To add a custom stopping condition, the user selects one of the RadioButtons, and enters in a value from the text box that appears subsequently. In the case of Target Accuracy, the user enters in a % value rather than a decimal value. For iterations, it would just be an integer value for the number of iterations they would like to stop at.

 ![Screen Shot 2020-07-16 at 4.33.19 PM.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/509/attachments/Screen%20Shot%202020-07-16%20at%204.33.19%20PM.png)

## Potential Issues
While training stops when the target accuracy is reached or the number of specified iterations have been completed, there are some issues with the visualizations. Take this image classifier I trained with early stopping, with a target accuracy of 80%:

![Screen Shot 2020-07-16 at 4.42.13 PM.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/509/attachments/Screen%20Shot%202020-07-16%20at%204.42.13%20PM.png)
![Screen Shot 2020-07-16 at 4.42.25 PM.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/509/attachments/Screen%20Shot%202020-07-16%20at%204.42.25%20PM.png)

As you can see from the screenshot of my terminal, 80% accuracy is reached on the 6th epoch and training is completed. But the result window shows that the accuracy reached is only 50%.

Related work items: #697