Merged PR 520: Autosuggest name for models
This PR adds autosuggested names for users when they click on the templates, in the `SelectModelModal`.

![image.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/520/attachments/image.png =400x)

Suggested names have the following format: `<template type with no spaces> <next highest number>`
For instance:
- `ObjectDetection 1` (if it's the first with the aforementioned format)
- `ImageClassification 3` (if there's an `ImageClassification 1` and `ImageClassification 2`)

Empty templates are called `Model <number>`.

Please note that the autosuggest is turned off as soon as the user has typed something in the name field. To reenable it, please close the modal and reopen it.

Related work items: #666