Merged PR 543: DataContainer Memory Usage Logger
This story creates the capability for the DataContainer to check/update how much memory it's utilizing. Currently, we're capping the maximum amount of memory that the DataContainer is allowed to be used at _**25% of the total memory**_.

The DataContainer will now display WARNINGS if it has reached/surpassed the allowed limit. In future iterations, the DataContainer will not be allowed to add any further experiments and must remove existing experiments to create space.

Please review the following files:
------------------------------------

**backend/perceptilabs/api/data_container.py**
- Added capability for DataContainer to track memory usage during processing incoming messages
- Added capability to remove experiment(s)

**backend/perceptilabs/api/test_data_container.py**
- Added additional testing to cover removal of experiment(s)

**backend/perceptilabs/dataschema/dc_memory_used.json**
- Added dataschema for DataContainer to use data_logger

**backend/perceptilabs/utils.py**
- Added functionality to allow measurement of incoming data objects (get_object_size)

As always, please let me know if there's any outstanding issues/concerns/questions that I should investigate :)

Related work items: #801