Merged PR 551: Adds ability to create GitHub issues
This PR adds a REST endpoint in **rygg** that allows for the creation of GitHub issues, as per #811.

To test, please:
- set the **GITHUB_API_KEY** env var to `01abf009e1725f5fac3350c30b3705c3fc1d9133`,
- set the **GITHUB_API_ENDPOINT** env var to `https://api.github.com/repos/dhplabs/issues-testground/issues`,
- create a **HTTP POST** request to `http://localhost:8000/issues/` with a JSON of the following format:

```
{
    "title": "<some title>",
    "body": "<some text>"
}
```

The response is a JSON with a `html_url` property (an example value is "https://github.com/dhplabs/issues-testground/issues/11").
Navigate to that link to check that the issue has been created as expected.

---

Please note, this is currently implemented with the use of Personal Access Tokens (PAT) and should be swapped out when we have our GitHub OAuth application in place.