Merged PR 607: Training Layers are runnable and throw errors before training starts now
# Introduction
This was a two part story for the following work item:
 #524

The first part involved refactoring all training layers so that they would be runnable, the second part involved creating a strategy in `lightweight2.py` for training layers so the lightweight core could respond to the errors when the layers are run.

# Part 1
To refactor the training layers, I separated everything preceding the training loop into a separate function called `init_layer()`. This function has two macros associated with variable initializations for distributed training and normal training for applicable training layers.

# Part 2
The strategy for the TrainingLayer is simple for now. I leverage the `instantiation_error` parameter in the `LayerResults` named tuple to pass any initialization errors caught by the strategy when running the `init_layer()` function. To see the code in more detail, see `lightweight2.py/TrainingStrategy`.

![image.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/607/attachments/image.png)

# Conclusion and Follow-up Questions
Currently, the frontend shows lingering error notifications under some circumstances when errors no longer exist. The following error is thrown:
`<LayerName> already exists, disallowed. Did you mean to set reuse=True or reuse=tf.AUTO_REUSE in VarScope?`

My initial thought was to do the canonical solution for this bug, which is to invoke `tf.reset_default_graph()`. However for reasons that are not clear to me, the following error is thrown when running integration tests:

`Do not use tf.reset_default_graph() to clear nested graphs. If you need a cleared graph, exit the nesting and create a new graph.`

This could potentially be because of multithreading on the kernel side.

**Side note**: the frontend (training and everything) runs well when invoking `tf.reset_default_graph()`, the integration tests are where it breaks.

Related work items: #524