Merged PR 638: Renamed tutorial data input & label files to be more intuitive
See `perceptilabs/backend/perceptilabs/tutorial_data` for changes.

Right now, the format is `<model_type>_<description>_<inputs/outputs>` for the file names.
An example of this is `classification_mnist_inputs.npy`. Not all files have a description because the model name is sufficient in those cases. For example, `linreg_inputs.npy` is one of those models in my opinion, feel free to give a description you think could be good for this situation in particular and others like it.

Related work items: #1014