Merged PR 662: Load Data replaces Choose File/Folder for DataData Env
Currently, we have two buttons, `Choose File` & `Choose Folder` for the users to utilize which option they want to pick. This PR allows for users to only need to choose `Load Data` , regardless of file/folder selection.

# Please Review the Following Files:
---------------------------------------
**/backend/perceptilabs/layers/datadata/imports.json**
- Added missing imports that were missing from `DataData`

**/frontend/src/renderer/components/different/file-picker.vue**
- Changed frontend to now utilize a `multimode` rather than `file` & `folder`

**/frontend/src/renderer/components/network-elements/elements/data-data/set-data-data.vue**
- Removed `Choose File` & `Choose Folder` buttons
- Added `Load Data` button

This work was done in joint with David Han, so credit to him as well for helping!

As always, please let me know if there are any concerns/comments/improvements. Thanks team! :)

Related work items: #1033