Merged PR 682: Fixing wrong line number in the lightweight core
The actual code shown to the user differs from the one we're actually running (the user doesn't see the import statements). So when an exception is thrown, the line number is off.

Since the exception is formatted by the traceback module I had  to resort to some regex magic to remove the difference between ran/shown code.

Related work items: #1002