Merged PR 688: Fix for missing frontend assets
This PR fixes the bug where the images in the "what's new" and the "create model" pages are missing.

The culprit was that the added files where not updated in the `package_data.json` file (used build by the build).

As a side effect, the `favicon.ico` file has also been added.