Merged PR 676: CSVs no longer stack columns in the preview
This is a PR for the bug #1037 .

# Files Changed
Changes were made by @<David Han> in the frontend for this one. See `set-data-data.vue` for those.

There were also changes made to the following files:
- `datadata/spec.py`: Added support for a tuple of strings as selected columns, and logic to deal with the situation in which a dictionary is passed to the kernel from the frontend rather than a list. A tuple of strings should be valid and is sometimes passed as the selected columns. This is also addressed in the next file where significant changes were made.
- `file_utils.j2`: Logic to check whether the selected columns are strings or integers, and additional logic to deal with the situation (which happened quite a bit when I was debugging) in which the string passed to the kernel was actually the index of the selected column wrapped inside a string. Whenever this happened, it was appended to a list of the selected columns with the correct information (so either the string names of the columns or the indexes as integers)

# Conclusion
These changes are rather convoluted, I must admit. A lot of these issues stem from the inconsistency in what's sent back from the frontend to the kernel though, I don't think the frontend should be sending dictionaries or strings at any point to the kernel given what's returned in the file David Han and I were originally looking at, `set-data-data.vue`. However, until we figure out why that's the case, I think the next best thing we can do is to look at workarounds for it within the kernel.

# Update 9/21/20
@<David Han> was able to figure out the root of the issue, and I've made some other small changes on the kernel to revert the workaround I had before. I also made some small changes to the frontend as well under David's supervision.

Related work items: #1037