Merged PR 588: Fix for bug 938: CTRL + Z not working
This PR fixes the issue where the undoing a delete via CTRL + Z not working.

This PR also allows the undo/redo of changes to:
- settings,
- layer positions,
- layer addition/removals.

![bug938_proof_of_fix.gif](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/588/attachments/bug938_proof_of_fix.gif)

Please note that there is an issue with the DataData layer, where all actions require 4x the undos of a normal layer.

---

**Update September 21:**
Merged in master and fixed the behaviour that was broken because of it.
![undoFix.gif](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/588/attachments/undoFix.gif)

Related work items: #938