Merged PR 687: Adds auto-close of checklist when all items are completed
This PR adds the auto-close of checklist when all items are completed.
The order doesn't matter and will happen when the last item is completed.

The checklist can be opened up again in the help panel at the top menu bar.
A related addition is the "Skip" button that hides the checklist:
![image.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/687/attachments/image.png =250x)