Merged PR 739: Fixes deleted models not showing up in Model Hub
This PR updates how the model is retrieved in efforts to solve the following bug:

**Models don't get "Deleted" properly:**
- Create a model in Model Hub
- Manually delete that model from your folder
- Refresh the browser

Since this affects what gets populated on app load, I would greatly appreciated it if you would take some time to test this fix.

Thanks for @<Anton Bourosu>  for pointing out that this bug has to do with a tricky watcher in `App.vue`.