Merged PR 816: [1103] Orders the networks by their model id
Page reloads or importing an older model with epoch date as the model id (instead of an id from rygg) occasionally causes the tab order to be unpredictable.

This PR implements the following changes:
- tabs are ordered by their model/network Id and
- model ids are now updated upon import.

This means that the order in Model Hub will always be the same as the order in the different views.

As a bonus, this PR also fixes a bug where imported models could not be deleted properly if its model id was a string (not deserialized properly).

This PR does not tackle the issue where tabs might not be selected upon page reload.

Related work items: #1103