Merged PR 855: Previews not showing if classification labels connection was unset at some point
Starting with a network like this (dense connected to predictions):

[Data] ---> [Dense] ---> [Classification]

For any layer added to this net, no previews were shown despite. The cause was that the autosettings module expected unset connections to be none, while the tf2 layers expected unset layers to be LayerConnections but with the layer IDs unset instead.

## Fix (implementation details)

Made connection_labels and connection_predictions default to None and added a getter for the tf2x layers to use.

Related work items: #1207