Merged PR 852: Force preview updates for layers that went through autosettings
Force preview updates for layers that went through autosettings

# Why the change was needed

Typically, the frontend specifies which layers it expects a new preview from in order to reduce bandwidth usage. This works for trivial cases, but not when the autosettings module has triggered an updated preview image. This means that sometimes the kernel should force a preview update even though the frontend does not expect it.

# Implementation details

Added a method to get layers that are different between the old and new graph. For the ones that are updated the Kernel will send a preview even though the frontend has not requested it.

Related work items: #1202