Merged PR 866: [1198] Tabs not created when importing a model
This PR fixes an issue where the newly imported network does not create a tab in the modeling workspace.
The user has to navigate to a different view and back to see the network.

The fix is implemented by setting the state of an **empty-navigation** component correctly to "0".

**Please note**: to reproduce, make sure that all of the networks are hidden or removed in the modeling workspace. It will not be reproducible if there is another tab active.

Tested with both local and GitHub networks.
The following link is so that you can save some time testing: https://github.com/dhplabs/ImageClassification-CNN-75.

Related work items: #1198, #1224