Merged PR 885: Expanded sentry coverage
Initializing sentry earlier in the kernels startup phase. Errors before a certain point are tagged with the tag "event-type" == "startup-error"

Note: removing the setting of the user from mainServer is safe since we're no longer passing the user id as a command line argument. Instead, it is set in mainInterface.py:542

Related work items: #1256