Merged PR 869: [story/906] Lock the compoennt if code was edited
Welcome to the description section of your PR!

This story implements the black overlay in front of the component settings.

To see that:
- Open any component code
- Edit code and save
The overlay should be on top of settings with a text message and lock icon

When pressing on reset component button it should reset the settings if it was changed and compoenent code.

Related work items: #906