Merged PR 881: Triggers PyPi check when tool loads
This PR adds a check for newer versions of the `perceptilabs` package on PyPi, when the app loads.

The original implementation called the check when the user logs in.
However, we are now using KeyCloak and the file containing the check is not longer used.

To test, please:
- change: `__version__ = "development" `to` __version__ = "0.9"` in the `rygg/__init__.py` file and
- reload the modeling tool.

---
Welcome to the description section of your PR!

Before submitting your PR, please make sure that the description contains:
- [X] Implementation details
- [x] Why the change was needed
- [x] (If applicable) Where to start reviewing and if something needs extra testing
- [ ] (If applicable) Performance changes

And that you:
- [ ] Include any new files in included_files.txt
- [x] Make sure you can build the wheel and run it
- [ ] Update any relevant Readme's

After the PR has gone through, please:
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack