Merged PR 880: Fixes corrupt checkpoint_model.json (synchronous frontend save)
When the users starts training a model:
- the frontend initiates a save of network (to `model.json`) asynchronously and
- the kernel copies the `model.json` to `checkpoint_model.json`.

The issue is when the first save action is slow or not finished, the created `checkpoint_model.json` might be incomplete and corrupt.

This PR makes the save operation initiated by the frontend synchronous.

This can be tested with large networks or by introducing an artificial delay in the `fileserver_saveModelJson` call.

----
Before submitting your PR, please make sure that the description contains:
- [x] Implementation details
- [x] Why the change was needed
- [ ] (If applicable) Where to start reviewing and if something needs extra testing
- [ ] (If applicable) Performance changes

And that you:
- [ ] Include any new files in included_files.txt
- [ ] Make sure you can build the wheel and run it
- [ ] Update any relevant Readme's

After the PR has gone through, please:
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack