Merged PR 942: Update dependency order to install correct numpy version
There has been a few reports from users where they have not been able to start the Kernel thanks to numpy crashing.
This is because the numpy version installed is 1.19.4 rather than the one we had in requirements (1.16.4).
To make sure that the correct numpy version gets installed, we adjust the order of the packages.

This PR also updates the version of the tool to 0.11.5 in all relevant places and changes the numpy version to 1.18.5 to avoid this issue: #1291.

As this is a release, please try to install the tool and make sure that the installation works as well as the base functionality of the tool. There are 9 PRs which has gone in since last release.
To test the new version you can install it through:
$ pip install pl-nightly==0.11.5.20201119.3

**Note:** Signing off on this PR is the same as signing off on the release, when this PR is merged we will push the release. Please don't merge anything else before this one has gone through.

Related work items: #1291