Merged PR 926: Use existing weights when starting training (tf2x)
# Parent branch

Based on [this PR](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/PerceptiLabs/pullrequest/925). Review using: git diff story/1146_export

# Implementation details:

* Added support for saving/loading weights in the training layer so that we can resume training where we left off.

# Files to review:

* **/backend/conftest.py**: made load_checkpoint an optional parameter in make_graph_spec
* **backend/perceptilabs/layers/trainclassification/tf2x_template.j2**: added support for loading weights via a checkpoint. Also added support for automatically saving weights.
* **backend/perceptilabs/layers/trainclassification/test_trainclassification_integration.py**: added tests and made load_checkpoint=True/False explicit

# Note:

Triggering a checkpoint load will give rise to the following warning:

_"Two checkpoint references resolved to different objects ({'output': <tf.Tensor 'prediction_model/deep_learning_fc__dense_1_keras/Identity_2:0' sh..."_

After turning over a good amount of stones I've realized that this warning is caused by the fact that the input tensors of a model are _initialized_ as non-eager tensors and are not converted to eager tensors until data has been passed through the model. Therefore:

* Since the saving happens post training, the model is saved using eager tensors, resulting in a checkpoint with concrete value for the model inputs.
* When we load the checkpoint, however, the model inputs are still non-eager tensors and therefore they don't expect any concrete values from the checkpoint.

Thus, the apparent inconsistency.

It seems like the TF team has been addressing similar issues recently (see [this](https://stackoverflow.com/questions/64403061/how-to-convert-a-tensor-to-eager-tensor-in-tensorflow-2-1-0) post). The warning is slightly more informative in full tensorflow 2.3.1, but the issue is still not resolved for our case. Ultimately, it _shouldn't_ have any practical consequences since we are only showing previews of layer outputs, but it might still cause some alarm for a user scanning the logs. Either way, please be extra thorough when testing the functionality.

**TLDR: It is my conclusion that this warning shouldn't have any practical consequences, but we should be lookout for issues both in the review process and after integrating**

Related work items: #1149