Merged PR 922: [1251] Adds user questionnaire
As a company, we want to get to know our customer better, so we're encouraging the users to fill in a questionnaire about their relationship with ML.

This PR add a questionnaire that shows up if the user has never filled it in or cancelled it before.

---
**Implementation:**
When the tool loads, we check to see if there's a "first login" attribute in the user token.
If the attribute is "true" or not present, we show the questionnaire.

The "first login" attribute is updated to `false` in KeyCloak when the user clicks on `Cancel`or `Confirm`.

If the user sends the questionnaire (by clicking on `Confirm`), two other attributes will be sent to KeyCloak as well:
- `questionnaire`
- `questionnaire-responses`

---
**To test:**

*** **Video of the steps below: [20201112_1251_questionnaire_test.mov](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/922/attachments/20201112_1251_questionnaire_test.mov)**

Just start the tool and the questionnaire should pop up.

To verify answers in KeyCloak, Rohan has helped to create temporary user accounts. :pray:
Please:
- navigate to https://keycloak.dev.perceptilabs.com:8443/auth/,
- click on _Administration Console_, and
- login using your Perceptilabs login with the password: _changeme_.

It should look like this in KeyCloak at the start:
![image.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/922/attachments/image.png =600x)

The attributes should look like this after submitting the questionnaire:
![image (2).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/922/attachments/image%20(2).png =600x)

---
**Extra things to check:**
- please make sure that the "What's New" page is showing up or hidden correctly

---
**Updates:**
---

**2020-11-17 update**

- Locked the 2nd page until there there is an answer on the 1st page.
- Added MixPanel tracking for the questionnaire responses.
![image (4).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/922/attachments/image%20(4).png =600x)

**2020-11-16 update**

- Locked the 3rd page until there there is an answer on the 2nd page.

Related work items: #1251