Merged PR 955: Merge layer output missing while training
When the inputs to Merge layer are Conv layers, Merge layer output has similar shape to that of Conv layers. For the visualization to work properly in the frontend, output has to be modified in the same way as for a conv layer.

Corresponding changes were made in coreInterface.py file.

Related work items: #1299