Merged PR 963: [1314] Adds sorting of files in the file chooser
Prior to this update, the modeling tool used the (unintuitive) sorting as it came from the file server service.
This PR adds sorting (case insensitive, by filenames) of files in the file chooser.

With this update, the `v-if` directive has been moved out of the block containing the `v-for` directive.
Not a problem in our case, but would save us some time if we ever start using linting:
https://github.com/vuejs/eslint-plugin-vue/blob/master/docs/rules/no-use-v-if-with-v-for.md

---

Before submitting your PR, please make sure that the description contains:
- [ ] Implementation details
- [x] Why the change was needed
- [ ] (If applicable) Where to start reviewing and if something needs extra testing
- [ ] (If applicable) Performance changes

And that you:
- [ ] Include any new files in included_files.txt
- [ ] Make sure you can build the wheel and run it
- [ ] Update any relevant Readme's

After the PR has gone through, please:
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #1314