Merged PR 964: [1050] Adds check before crossing off checklist item in "Get started checklist"
The 2nd option in the "Get started checklist" is "Start training your model", whose state would be updated to done as soon as training is started. This happens even for models with errors, where a training session hasn't really happened.

This PR updates the "Start training your model" status when training a model with no errors.
An alternative approach would be to update the value in the `getTrainingStatistics` action, but I feel that updating the status there is unnecessary for every dispatch of the action.

**Note**: easiest to test in an incognito window because of potentially saved tutorial progress.

Network with errors:
![20201127_bug1050_networkWithErrors.gif](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/964/attachments/20201127_bug1050_networkWithErrors.gif)

Network without errors:
![20201127_bug1050_networkWithoutErrors.gif](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/964/attachments/20201127_bug1050_networkWithoutErrors.gif)

---

Before submitting your PR, please make sure that the description contains:
- [ ] Implementation details
- [x] Why the change was needed
- [ ] (If applicable) Where to start reviewing and if something needs extra testing
- [ ] (If applicable) Performance changes

And that you:
- [ ] Include any new files in included_files.txt
- [ ] Make sure you can build the wheel and run it
- [ ] Update any relevant Readme's

After the PR has gone through, please:
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Adds check before crossing off checklist item

Related work items: #1050