Merged PR 1003: [993] Adds missing headless internal state
When working with the frontend, the setting is headless state is much simpler when the previous state doesn't need to be taking into consideration.

This PR simplifies the handling of headless state for the frontend.

Kudos to @Derek Ly for working on the work kernel-side 🙏
In the kernel, logic to handle headless state changes has been added.

---

Before submitting your PR, please make sure that the description contains:
- [ ] Implementation details
- [x] Why the change was needed
- [ ] (If applicable) Where to start reviewing and if something needs extra testing
- [ ] (If applicable) Performance changes

And that you:
- [ ] Include any new files in included_files.txt
- [ ] Make sure you can build the wheel and run it
- [ ] Update any relevant Readme's

After the PR has gone through, please:
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #993