Merged PR 1045: Created TF2x implementation for Switch
## Implementation Details
Similar to adding any other TF2x component, appropriate import files, test cases, additions to `definitions.py`, and Jinja template were created. Because the switch component is very simple, the implementation isn't different logically compared to the TF1x implementation.

Something that should be noted is that running the layer looks like this in the jinja template:

<pre><code>
        {% call indented_if(layer_spec.selected_var_name == 'input1')%}
           y = inputs['input1']
        {% endcall %}

        {% call indented_if(layer_spec.selected_var_name == 'input2')%}
            y = inputs['input2']
        {% endcall %}
</code></pre>

The reason I've implemented it this way is because passing in `layer_spec.selected_var_name` directly as the name of the dictionary key (like this: `y = inputs[{{layer_spec.selected_var_name}}]`) results in an error, this seems to be a quirk with Jinja because the conditionals are entered properly.

## Where to Start Reviewing
* `/backend/perceptilabs/layers/mathswitch/tf2x_template.j2`: Contains the actual layer implementation.
* `/backend/perceptilabs/layers/mathswitch/tf2x_imports.json`: Imports for tf2x implementation.
* `/backend/perceptilabs/layers/mathswitch/test_mathswitch_integration.py`: Added a test case for the tf2x implementation.

Related work items: #1481